requirements.txt
streamlit==1.35.0
langchain>=0.1.14
langchain-core>=0.1.45
langchain-community>=0.0.36
langchain-mongodb>=0.1.0
ollama>=0.1.0
pymongo==4.6.0
huggingface-hub==0.21.4
sentence-transformers==2.3.1
python-dotenv==1.0.0
langsmith==0.1.62
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.4
langchain-openai
